# docs/api/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: Document Intelligence API
  description: |
    **AI-powered document analysis and processing API**

    This API provides comprehensive document intelligence capabilities including:
    - Document upload and processing
    - Named Entity Recognition (NER)
    - Document classification
    - Relation extraction
    - Knowledge graph generation
    - Batch processing

    ## Authentication

    API key authentication is optional. When enabled, include the API key in the header:
    ```
    X-API-Key: your-api-key-here
    ```

    ## Rate Limiting

    - 100 requests per minute per IP (default)
    - 1000 requests per hour per API key

    ## Getting Started

    1. Start the API server: `python doc-api-server.py`
    2. Access Swagger UI: http://localhost:5000/docs
    3. Try the health check endpoint: `GET /api/v1/health`

  version: 1.0.0
  contact:
    name: Document Management System
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: http://localhost:8000/api/v1
    description: Alternative development server
  - url: https://api.example.com/v1
    description: Production server
...
paths:
...
  /documents:
    post:
      tags:
        - Documents
      summary: Upload and process document
      description: |
        Upload a document for processing. Supported formats:
        - PDF (.pdf)
        - Word (.docx)
        - Text (.txt)
        - HTML (.html)
        - Markdown (.md)

        The API will:
        1. Parse the document
        2. Extract text content
        3. Perform NER (Named Entity Recognition)
        4. Classify the document
        5. Extract relations
        6. Build a knowledge graph (optional)
      operationId: uploadDocument
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file to process
                build_graph:
                  type: boolean
                  default: false
                  description: Whether to build knowledge graph
                extract_relations:
                  type: boolean
                  default: true
                  description: Whether to extract relations
      responses:
        '200':
          description: Document processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
              examples:
                success:
                  value:
                    document_id: "550e8400-e29b-41d4-a716-446655440000"
                    filename: "contract.pdf"
                    processed_at: "2026-01-14T12:00:00Z"
                    statistics:
                      pages: 5
                      words: 1523
                      characters: 9842
                    classification:
                      category: "contract"
                      confidence: 0.95
                      probabilities:
                        contract: 0.95
                        invoice: 0.03
                        report: 0.02
                    entities:
                      - text: "John Doe"
                        type: "PERSON"
                        start: 45
                        end: 53
                        confidence: 0.98
                      - text: "Acme Corp"
                        type: "ORG"
                        start: 120
                        end: 129
                        confidence: 0.96
                    relations:
                      - source: "John Doe"
                        source_type: "PERSON"
                        relation: "works_for"
                        target: "Acme Corp"
                        target_type: "ORG"
                        confidence: 0.89
                    knowledge_graph: null
        '400':
          description: Invalid request (e.g., unsupported file format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (invalid API key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large (max 50MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /extract/entities:
    post:
      tags:
        - Extraction
      summary: Extract entities from text
      description: |
        Extract named entities (NER) from input text.

        Supported entity types:
        - PERSON: People names
        - ORG: Organizations
        - LOC: Locations
        - DATE: Dates
        - MONEY: Monetary values
        - PERCENT: Percentages
        - TIME: Times
        - GPE: Geopolitical entities
      operationId: extractEntities
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextInput'
            examples:
              example1:
                summary: Contract text
                value:
                  text: "John Doe from Acme Corp signed the contract on January 15, 2026 for $50,000."
                  options:
                    min_confidence: 0.7
      responses:
        '200':
          description: Entities extracted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  entities:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityResponse'
                  count:
                    type: integer
                    example: 5
              examples:
                success:
                  value:
                    entities:
                      - text: "John Doe"
                        type: "PERSON"
                        start: 0
                        end: 8
                        confidence: 0.98
                      - text: "Acme Corp"
                        type: "ORG"
                        start: 14
                        end: 23
                        confidence: 0.96
                      - text: "January 15, 2026"
                        type: "DATE"
                        start: 49
                        end: 65
                        confidence: 0.99
                      - text: "$50,000"
                        type: "MONEY"
                        start: 70
                        end: 77
                        confidence: 0.97
                    count: 4
  /extract/relations:
    post:
      tags:
        - Extraction
      summary: Extract relations from text
      description: |
        Extract semantic relations between entities in text.

        Supported relation types:
        - works_for
        - located_in
        - part_of
        - owned_by
        - manages
        - affiliated_with
      operationId: extractRelations
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextInput'
            examples:
              example1:
                summary: Business text
                value:
                  text: "John Doe is the CEO of Acme Corp, headquartered in New York."
                  options:
                    min_confidence: 0.6
      responses:
        '200':
          description: Relations extracted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  relations:
                    type: array
                    items:
                      $ref: '#/components/schemas/RelationResponse'
                  count:
                    type: integer
              examples:
                success:
                  value:
                    relations:
                      - source: "John Doe"
                        source_type: "PERSON"
                        relation: "works_for"
                        target: "Acme Corp"
                        target_type: "ORG"
                        confidence: 0.92
                      - source: "Acme Corp"
                        source_type: "ORG"
                        relation: "located_in"
                        target: "New York"
                        target_type: "GPE"
                        confidence: 0.88
                    count: 2
  /classify:
    post:
      tags:
        - Classification
      summary: Classify document
      description: |
        Classify document into predefined categories.

        Supported categories:
        - contract
        - invoice
        - report
        - email
        - letter
        - form
        - other
      operationId: classifyDocument
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextInput'
            examples:
              contract:
                summary: Contract text
                value:
                  text: "This agreement is made between Party A and Party B..."
      responses:
        '200':
          description: Document classified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationResponse'
              examples:
                success:
                  value:
                    category: "contract"
                    confidence: 0.94
                    probabilities:
                      contract: 0.94
                      letter: 0.03
                      form: 0.02
                      other: 0.01
  /graph/build:
    post:
      tags:
        - Knowledge Graph
      summary: Build knowledge graph
      description: |
        Build a knowledge graph from text, extracting entities and their relationships.

        Output formats:
        - json: JSON graph structure
        - networkx: NetworkX-compatible format
        - cypher: Cypher queries for Neo4j
      operationId: buildKnowledgeGraph
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TextInput'
                - type: object
                  properties:
                    format:
                      type: string
                      enum: [json, networkx, cypher]
                      default: json
            examples:
              example1:
                value:
                  text: "John Doe works for Acme Corp in New York. Jane Smith is the CFO."
                  format: "json"
      responses:
        '200':
          description: Knowledge graph built successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        label:
                          type: string
                        type:
                          type: string
                  edges:
                    type: array
                    items:
                      type: object
                      properties:
                        source:
                          type: string
                        target:
                          type: string
                        relation:
                          type: string
                  statistics:
                    type: object
                    properties:
                      node_count:
                        type: integer
                      edge_count:
                        type: integer
  /batch/process:
    post:
      tags:
        - Batch
      summary: Batch process documents
      description: |
        Process multiple documents in a batch job.
        Upload up to 100 documents at once.
      operationId: batchProcess
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Documents to process (max 100)
                build_graph:
                  type: boolean
                  default: false
      responses:
        '202':
          description: Batch job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
              examples:
                created:
                  value:
                    job_id: "batch-123e4567-e89b-12d3-a456-426614174000"
                    status: "pending"
                    total_documents: 25
                    processed: 0
                    failed: 0
                    created_at: "2026-01-14T12:00:00Z"
                    updated_at: "2026-01-14T12:00:00Z"
  /batch/{job_id}:
    get:
      tags:
        - Batch
      summary: Get batch job status
      description: Check the status of a batch processing job
      operationId: getBatchJob
      security:
        - ApiKeyAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          description: Batch job ID
          schema:
            type: string
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
              examples:
                processing:
                  value:
                    job_id: "batch-123e4567-e89b-12d3-a456-426614174000"
                    status: "processing"
                    total_documents: 25
                    processed: 15
                    failed: 0
                    created_at: "2026-01-14T12:00:00Z"
                    updated_at: "2026-01-14T12:05:30Z"
  /stats:
    get:
      tags:
        - Statistics
      summary: Get API statistics
      description: Retrieve API usage statistics
      operationId: getStatistics
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_requests:
                    type: integer
                    example: 1523
                  documents_processed:
                    type: integer
                    example: 342
                  entities_extracted:
                    type: integer
                    example: 8456
                  relations_extracted:
                    type: integer
                    example: 1234
                  uptime:
                    type: number
                    format: float
                    example: 3600.5
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional in development mode)
  schemas:
    TextInput:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Input text to process
          minLength: 1
          maxLength: 100000
          example: "John Doe works for Acme Corp."
        options:
          type: object
          description: Processing options
          additionalProperties: true
          example:
            min_confidence: 0.7
            language: "en"
...
    DocumentResponse:
      type: object
      required:
        - document_id
        - filename
        - processed_at
        - statistics
        - classification
        - entities
        - relations
      properties:
        document_id:
          type: string
          format: uuid
          description: Document UUID
          example: "550e8400-e29b-41d4-a716-446655440000"
        filename:
          type: string
          description: Original filename
          example: "contract.pdf"
        processed_at:
          type: string
          format: date-time
          description: Processing timestamp
          example: "2026-01-14T12:00:00Z"
        statistics:
          type: object
          description: Document statistics
          properties:
            pages:
              type: integer
            words:
              type: integer
            characters:
              type: integer
        classification:
          $ref: '#/components/schemas/ClassificationResponse'
        entities:
          type: array
          items:
            $ref: '#/components/schemas/EntityResponse'
        relations:
          type: array
          items:
            $ref: '#/components/schemas/RelationResponse'
        knowledge_graph:
          type: object
          nullable: true
          description: Knowledge graph (if requested)
...